# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Roomly household management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /households:
    get:
      operationId: listHouseholds
      tags: [households]
      summary: List all households
      responses:
        "200":
          description: List of households
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Household"
    post:
      operationId: createHousehold
      tags: [households]
      summary: Create a household
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HouseholdInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Household"
  /households/{householdId}:
    get:
      operationId: getHousehold
      tags: [households]
      summary: Get a household
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Household
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Household"
        "404":
          description: Not found
    patch:
      operationId: updateHousehold
      tags: [households]
      summary: Update a household
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HouseholdUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Household"

  # ── Members ────────────────────────────────────────────────────────────────
  /households/{householdId}/members:
    get:
      operationId: listMembers
      tags: [members]
      summary: List household members
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
    post:
      operationId: addMember
      tags: [members]
      summary: Add a member
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /members/{memberId}:
    patch:
      operationId: updateMember
      tags: [members]
      summary: Update a member
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
    delete:
      operationId: removeMember
      tags: [members]
      summary: Remove a member
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  # ── Rooms ─────────────────────────────────────────────────────────────────
...
  /households/{householdId}/expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List household expenses
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Log a shared expense
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{expenseId}:
    get:
      operationId: getExpense
      tags: [expenses]
      summary: Get an expense
      parameters:
        - name: expenseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    patch:
      operationId: updateExpense
      tags: [expenses]
      summary: Update an expense
      parameters:
        - name: expenseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense
      parameters:
        - name: expenseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /expenses/{expenseId}/settle:
    post:
      operationId: settleExpense
      tags: [expenses]
      summary: Mark an expense as settled
      parameters:
        - name: expenseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Settled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"

  # ── Chores ────────────────────────────────────────────────────────────────
...
  /landlord/households/{householdId}/payment-stats:
    get:
      operationId: getPaymentStats
      tags: [landlord]
      summary: Get on-time payment stats for a household
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentStats"
components:
  schemas:
...
    Household:
      type: object
      required: [id, name, address, totalRent, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: string
        totalRent:
          type: number
        leaseStart:
          type: ["string", "null"]
        leaseEnd:
          type: ["string", "null"]
        landlordName:
          type: ["string", "null"]
        landlordEmail:
          type: ["string", "null"]
        createdAt:
          type: string
    HouseholdInput:
      type: object
      required: [name, address, totalRent]
      properties:
        name:
          type: string
          minLength: 1
        address:
          type: string
          minLength: 1
        totalRent:
          type: number
          minimum: 0
        leaseStart:
          type: string
        leaseEnd:
          type: string
        landlordName:
          type: string
        landlordEmail:
          type: string
    HouseholdUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        address:
          type: string
        totalRent:
          type: number
          minimum: 0
        leaseStart:
          type: string
        leaseEnd:
          type: string
        landlordName:
          type: string
        landlordEmail:
          type: string

    # ── Member ────────────────────────────────────────────────────────────
    Member:
      type: object
      required: [id, householdId, name, email, role, rentShare, moveInDate, createdAt]
      properties:
        id:
          type: integer
        householdId:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, member]
        rentShare:
          type: number
        avatarColor:
          type: ["string", "null"]
        roomId:
          type: ["integer", "null"]
        moveInDate:
          type: string
        createdAt:
          type: string
    MemberInput:
      type: object
      required: [name, email, rentShare]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        role:
          type: string
          enum: [admin, member]
          default: member
        rentShare:
          type: number
          minimum: 0
        avatarColor:
          type: string
        roomId:
          type: integer
        moveInDate:
          type: string
    MemberUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, member]
        rentShare:
          type: number
          minimum: 0
        avatarColor:
          type: string
        roomId:
          type: integer

    # ── Room ──────────────────────────────────────────────────────────────
...
    Expense:
      type: object
      required: [id, householdId, title, amount, category, paidByMemberId, splitType, settled, date, createdAt]
      properties:
        id:
          type: integer
        householdId:
          type: integer
        title:
          type: string
        amount:
          type: number
        category:
          type: string
          enum: [groceries, utilities, internet, streaming, household, transport, other]
        paidByMemberId:
          type: integer
        splitType:
          type: string
          enum: [equal, custom]
        settled:
          type: boolean
        notes:
          type: ["string", "null"]
        date:
          type: string
        createdAt:
          type: string
        splits:
          type: array
          items:
            $ref: "#/components/schemas/ExpenseSplit"
    ExpenseSplit:
      type: object
      required: [id, expenseId, memberId, amount, paid]
      properties:
        id:
          type: integer
        expenseId:
          type: integer
        memberId:
          type: integer
        amount:
          type: number
        paid:
          type: boolean
    ExpenseInput:
      type: object
      required: [title, amount, category, paidByMemberId, splitType, date]
      properties:
        title:
          type: string
          minLength: 1
        amount:
          type: number
          minimum: 0
        category:
          type: string
          enum: [groceries, utilities, internet, streaming, household, transport, other]
        paidByMemberId:
          type: integer
        splitType:
          type: string
          enum: [equal, custom]
        notes:
          type: string
        date:
          type: string
        customSplits:
          type: array
          items:
            $ref: "#/components/schemas/CustomSplitItem"
    CustomSplitItem:
      type: object
      required: [memberId, amount]
      properties:
        memberId:
          type: integer
        amount:
          type: number
    ExpenseUpdate:
      type: object
      properties:
        title:
          type: string
        amount:
          type: number
          minimum: 0
        category:
          type: string
          enum: [groceries, utilities, internet, streaming, household, transport, other]
        notes:
          type: string
        date:
          type: string

    # ── Chore ─────────────────────────────────────────────────────────────
...
    PaymentStats:
      type: object
      required: [householdId, totalPayments, onTimeCount, lateCount, onTimeRate, averageDaysLate, monthlyBreakdown]
      properties:
        householdId:
          type: integer
        totalPayments:
          type: integer
        onTimeCount:
          type: integer
        lateCount:
          type: integer
        onTimeRate:
          type: number
        averageDaysLate:
          type: number
        monthlyBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/MonthlyPaymentSummary"
    MonthlyPaymentSummary:
      type: object
      required: [month, totalDue, totalPaid, onTime, late]
      properties:
        month:
          type: string
        totalDue:
          type: number
        totalPaid:
          type: number
        onTime:
          type: integer
        late:
          type: integer