# lib/api-zod/src/generated/types/expenseUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Roomly household management API
 * OpenAPI spec version: 0.1.0
 */
import type { ExpenseUpdateCategory } from './expenseUpdateCategory';

export interface ExpenseUpdate {
  title?: string;
  /** @minimum 0 */
  amount?: number;
  category?: ExpenseUpdateCategory;
  notes?: string;
  date?: string;
}