# lib/api-zod/src/generated/types/memberUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Roomly household management API
 * OpenAPI spec version: 0.1.0
 */
import type { MemberUpdateRole } from './memberUpdateRole';

export interface MemberUpdate {
  name?: string;
  email?: string;
  role?: MemberUpdateRole;
  /** @minimum 0 */
  rentShare?: number;
  avatarColor?: string;
  roomId?: number;
}