# lib/api-zod/src/generated/types/paymentStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Roomly household management API
 * OpenAPI spec version: 0.1.0
 */
import type { MonthlyPaymentSummary } from './monthlyPaymentSummary';

export interface PaymentStats {
  householdId: number;
  totalPayments: number;
  onTimeCount: number;
  lateCount: number;
  onTimeRate: number;
  averageDaysLate: number;
  monthlyBreakdown: MonthlyPaymentSummary[];
}